chore: Extracted upload queue processing code into its own action in order to reuse it elsewhere
