+++ src/drive/actions/index.jsx
@@ -5,6 +5,7 @@ import {
   getLoadedFilesCount,
   getLoadedFoldersCount
 } from '../reducers'
+import React from 'react'
 import { isCordova } from '../mobile/lib/device'
 import {
   saveFileWithCordova,
@@ -20,6 +21,9 @@ import {
 } from '../ducks/files/files'
 import { revokeLink as revokeSharingLink } from 'cozy-client'
 import * as availableOffline from '../ducks/files/availableOffline'
+import { alert } from '../lib/confirm'
+import { alertShow } from 'cozy-ui/react/Alerter'
+import QuotaAlert from '../components/QuotaAlert'
 
 import { ROOT_DIR_ID, TRASH_DIR_ID } from '../constants/config.js'
 
@@ -57,7 +61,9 @@ export const OPEN_FILE_E_NO_APP = 'OPEN_FILE_E_NO_APP'
 export const getOpenedFolderId = state => state.view.openedFolderId
 
 const HTTP_CODE_CONFLICT = 409
+const ALERT_LEVEL_INFO = 'info'
 const ALERT_LEVEL_ERROR = 'error'
+const ALERT_LEVEL_SUCCESS = 'success'
 
 export const META_DEFAULTS = {
   cancelSelection: true,
@@ -233,6 +239,31 @@ export const uploadedFile = file => {
       currentSort: getSort(getState())
     })
   }
+}
+
+export const uploadQueueProcessed = (loaded, quotas, conflicts, errors, t) => {
+  let action = { type: '' } // dummy action, we only use it to trigger an alert notification
+
+  if (quotas.length > 0) {
+    // quota errors have their own modal instead of a notification
+    alert(<QuotaAlert t={t} />)
+  } else if (conflicts.length > 0) {
+    action.alert = alertShow(
+      'upload.alert.success_conflicts',
+      { smart_count: loaded.length, conflictNumber: conflicts.length },
+      ALERT_LEVEL_INFO
+    )
+  } else if (errors.length > 0) {
+    action.alert = alertShow('upload.alert.errors', null, ALERT_LEVEL_ERROR)
+  } else {
+    action.alert = alertShow(
+      'upload.alert.success',
+      { smart_count: loaded.length },
+      ALERT_LEVEL_SUCCESS
+    )
+  }
+
+  return action
 }
 
 export const abortAddFolder = accidental => {
